simplify SP/Resource Browser: extract format helpers, stat types, split StatsTab
